refactor(lex-analysis): split diagnostics.rs by diagnostic category
